Added vCards to the message events.
This update ships a meta data addition to the Sidekiq XML event message.
From now on vCard data will be added for the sender and receiver(s) of
a single message. This allows third party applications to easily
identify the users without additional XMPP queries. When no vCard is
available for a user then the +<vCard>+ element is not added to its
respective +<from>+ and +<to>+ element.

Signed-off-by: Hermann Mayer <hermann.mayer92@gmail.com>